Update to New Move Algorithm
