--- easyGame/MainModule/MovePhysicsServise/ExtAntiHeroMovePhysicsServise.swift
@@ -1,9 +1,0 @@
-//
-//  ExtAntiHeroMovePhysicsServise.swift
-//  easyGame
-//
-//  Created by chulep on 17.02.2022.
-//
-
-import Foundation
-
